// src/app/interfaces/IPuntajeDialogData.ts
export interface IPuntajeDialogData {
  idPuntaje: number | null;
  idAlumno: number;
  nombreAlumno: string;
  apellidoAlumno: string;
  idProfesor: number;
  nombreProfesor: string;
  apellidoProfesor: string;
  idTema: number;
  tema: string;
  interes: number;
  complejidad: number;
  entendimiento: number;
  valoracion: number;
  observaciones: string;
}